Users edit events
